zephyr: Move the sync::Mutex into it's own file
Avoid cluttering the top-level sync module with specifics.
